+++ monkey/infection_monkey/credential_collectors/__init__.py
@@ -1,6 +1,4 @@
-from .i_credential_collector import ICredentialCollector
 from .credential_components.nt_hash import NTHash
 from .credential_components.lm_hash import LMHash
 from .credential_components.password import Password
 from .credential_components.username import Username
-from .credentials import Credentials
