+++ monkey/infection_monkey/i_puppet/__init__.py
@@ -10,3 +10,9 @@ from .i_puppet import (
     UnknownPluginError,
 )
 from .i_fingerprinter import IFingerprinter
+from .credential_collection import (
+    Credentials,
+    CredentialType,
+    ICredentialCollector,
+    ICredentialComponent,
+)
