+++ monkey/infection_monkey/i_puppet/credential_collection/__init__.py
@@ -1,0 +1,4 @@
+from .i_credential_collector import ICredentialCollector
+from .credentials import Credentials
+from .i_credential_component import ICredentialComponent
+from .credential_type import CredentialType
