+++ monkey/infection_monkey/i_puppet/credential_collection/credential_type.py
