+++ monkey/infection_monkey/i_puppet/credential_collection/credentials.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import Tuple
 
-from .credential_components.i_credential_component import ICredentialComponent
+from .i_credential_component import ICredentialComponent
 
 
 @dataclass(frozen=True)
