+++ monkey/infection_monkey/i_puppet/credential_collection/i_credential_collector.py
