+++ monkey/infection_monkey/i_puppet/credential_collection/i_credential_component.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from infection_monkey.credential_collectors.credential_type import CredentialType
+from .credential_type import CredentialType
 
 
 class ICredentialComponent(ABC):
